remove share dashboard component from shared library
