+++ projects/back-office/src/app/dashboard/pages/dashboard/dashboard.component.ts
@@ -336,17 +336,11 @@ export class DashboardComponent implements OnInit, OnDestroy {
     }
   }
 
-  /*  Display the ShareUrl modal with the route to access the dashboard.
+  /**
+   * Display the ShareUrl modal with the route to access the dashboard.
    */
   public onShare(): void {
-    let url = '';
-    this.route.params.subscribe((params: any) => {
-      url =
-        window.location.href.split('/applications')[0].split('//')[1] +
-        '/share/' +
-        params.id;
-    });
-
+    const url = `${window.origin}/share/${this.dashboard?.id}`;
     const dialogRef = this.dialog.open(ShareUrlComponent, {
       data: {
         url,
