+++ projects/back-office/src/app/share/share.component.html
@@ -1,1 +1,1 @@
-<safe-share-redirect></safe-share-redirect>
+<mat-spinner></mat-spinner>
