+++ projects/back-office/src/app/share/share.component.scss
@@ -1,0 +1,8 @@
+mat-spinner {
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  margin: auto;
+}
