+++ projects/back-office/src/app/share/share.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ShareComponent } from './share.component';
 
-describe('PreviewToolbarComponent', () => {
+describe('ShareComponent', () => {
   let component: ShareComponent;
   let fixture: ComponentFixture<ShareComponent>;
