+++ projects/back-office/src/app/share/share.module.ts
@@ -2,10 +2,13 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ShareComponent } from './share.component';
 import { ShareRoutingModule } from './share-routing.module';
-import { SafeShareRedirectModule } from '@safe/builder';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 
+/**
+ * Share Url access module.
+ */
 @NgModule({
   declarations: [ShareComponent],
-  imports: [CommonModule, ShareRoutingModule, SafeShareRedirectModule],
+  imports: [CommonModule, ShareRoutingModule, MatProgressSpinnerModule],
 })
 export class ShareModule {}
