+++ projects/back-office/src/environments/environment.ts
@@ -28,10 +28,10 @@ const authConfig: AuthConfig = {
  */
 export const environment = {
   production: false,
-  // apiUrl: 'https://oort-dev.oortcloud.tech/api',
-  // subscriptionApiUrl: 'wss://oort-dev.oortcloud.tech/api',
-  apiUrl: 'http://localhost:3000',
-  subscriptionApiUrl: 'ws://localhost:3000',
+  apiUrl: 'https://oort-dev.oortcloud.tech/api',
+  subscriptionApiUrl: 'wss://oort-dev.oortcloud.tech/api',
+  // apiUrl: 'http://localhost:3000',
+  // subscriptionApiUrl: 'ws://localhost:3000',
   frontOfficeUri: 'http://localhost:4200/',
   backOfficeUri: 'http://localhost:4200/',
   module: 'backoffice',
