--- projects/safe/src/lib/components/share-redirect/public-api.ts
@@ -1,2 +1,0 @@
-export * from './share-redirect.component';
-export * from './share-redirect.module';
