--- projects/safe/src/lib/components/share-redirect/share-redirect.component.html
@@ -1,1 +1,0 @@
-<mat-spinner></mat-spinner>
