--- projects/safe/src/lib/components/share-redirect/share-redirect.component.scss
@@ -1,8 +1,0 @@
-mat-spinner {
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    margin: auto;
-}
