--- projects/safe/src/lib/components/share-redirect/share-redirect.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { SafeShareRedirectComponent } from './share-redirect.component';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-
-@NgModule({
-  declarations: [SafeShareRedirectComponent],
-  imports: [MatProgressSpinnerModule],
-  exports: [SafeShareRedirectComponent],
-})
-export class SafeShareRedirectModule {}
